requirements.txt
addict==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.2.0
archspec @ file:///croot/archspec_1697725767277/work
argcomplete==3.2.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boltons @ file:///work/ci_py311/boltons_1677685195580/work
boto3==1.33.13
botocore==1.33.13
Brotli==1.1.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm==0.2.2
conda @ file:///croot/conda_1699303130622/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1700150901600/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
ConfigArgParse==1.7
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
dacite==1.7.0
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.5.0
einops==0.8.0
executing==2.0.1
fastapi==0.112.1
fastjsonschema==2.20.0
ffmpy==0.4.0
fiftyone==0.23.4
fiftyone-brain==0.16.0
fiftyone-db==1.1.1
fiftyone-desktop==0.33.3
filelock==3.15.4
Flask==3.0.3
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.6.1
ftfy==6.1.3
future==0.18.3
glob2==0.7
gradio==4.41.0
gradio_client==1.3.0
graphql-core==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.24.5
humanize==4.9.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna @ file:///work/ci_py311/idna_1676822698822/work
imageio==2.33.1
importlib_metadata==8.0.0
importlib_resources==6.4.3
inflate64==1.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.25
jsonlines==4.0.0
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.3
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mongoengine==0.24.2
motor==3.3.2
mpmath==1.3.0
multivolumefile==0.2.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.1
open3d==0.18.0
opencv-python==4.10.0.84
opencv-python-headless==4.9.0.80
openvino==2024.3.0
openvino-dev==2024.3.0
openvino-telemetry==2024.1.0
orjson==3.10.7
overrides==7.7.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
pexel_downloader==0.3.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plotly==5.18.0
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
pprintpp==0.4.0
priority==2.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.8
pybcj==1.0.2
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pymongo==4.6.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyppmd==1.1.0
pyquaternion==0.9.9
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.1.0
pyzstd==0.15.9
rarfile==4.1
referencing==0.35.1
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
ruff==0.6.1
s3transfer==0.8.2
safetensors==0.4.4
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.38.2
strawberry-graphql==0.138.1
sympy==1.13.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==1.0.8
tinycss2==1.3.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.14.3
transformers==4.44.0
triton==3.0.0
truststore @ file:///croot/truststore_1695244293384/work
typer==0.12.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
voxel51-eta==0.12.4
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.19.2
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
